requirements.txt
pyyaml==6.0.2
google-api-python-client==2.140.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
PyMySQL==1.1.1
beautifulsoup4==4.12.0
lxml==5.2.2
validictory==1.0.0